Add login validation
