+++ content/gameplay/transmissions/tx_evacuation.chunks.json
@@ -1,20 +1,56 @@
 [
   {
-    "chunkId": "tx_evacuation_01",
+    "chunkId": "tx_evacuation_01a",
     "transmissionId": "tx_evacuation",
     "lineId": "evacuation_line_01",
     "startMs": 0,
-    "endMs": 4000,
+    "endMs": 1400,
     "speakerLabel": "CIVIL-DEFENSE",
-    "text": "緊急警報。第4居住区パノラマドームにて大規模な磁気崩壊が観測されました。全市民は直ちに地下シェルターへ避難してください。"
+    "text": "緊急警報。"
   },
   {
-    "chunkId": "tx_evacuation_02",
+    "chunkId": "tx_evacuation_01b",
+    "transmissionId": "tx_evacuation",
+    "lineId": "evacuation_line_01",
+    "startMs": 1400,
+    "endMs": 2800,
+    "speakerLabel": "CIVIL-DEFENSE",
+    "text": "第4居住区パノラマドームにて大規模な磁気崩壊が観測されました。"
+  },
+  {
+    "chunkId": "tx_evacuation_01c",
+    "transmissionId": "tx_evacuation",
+    "lineId": "evacuation_line_01",
+    "startMs": 2800,
+    "endMs": 4000,
+    "speakerLabel": "CIVIL-DEFENSE",
+    "text": "全市民は直ちに地下シェルターへ避難してください。"
+  },
+  {
+    "chunkId": "tx_evacuation_02a",
     "transmissionId": "tx_evacuation",
     "lineId": "evacuation_line_02",
     "startMs": 4500,
+    "endMs": 6000,
+    "speakerLabel": "CIVIL-DEFENSE",
+    "text": "繰り返します……市民の皆さんは直ちに避難を……。"
+  },
+  {
+    "chunkId": "tx_evacuation_02b",
+    "transmissionId": "tx_evacuation",
+    "lineId": "evacuation_line_02",
+    "startMs": 6000,
+    "endMs": 7200,
+    "speakerLabel": "CIVIL-DEFENSE",
+    "text": "救援部隊の到着は……絶望的……"
+  },
+  {
+    "chunkId": "tx_evacuation_02c",
+    "transmissionId": "tx_evacuation",
+    "lineId": "evacuation_line_02",
+    "startMs": 7200,
     "endMs": 8500,
     "speakerLabel": "CIVIL-DEFENSE",
-    "text": "繰り返します……市民の皆さんは直ちに避難を……。救援部隊の到着は……絶望的……自力で……"
+    "text": "自力で……"
   }
 ]
