+++ content/gameplay/transmissions/tx_good_morning.json
@@ -13,15 +13,25 @@
     "reboot"
   ],
   "transcriptChunkIds": [
-    "chunk_gm_001",
-    "chunk_gm_002",
-    "chunk_gm_003",
-    "chunk_gm_004",
-    "chunk_gm_005",
-    "chunk_gm_006",
-    "chunk_gm_007",
-    "chunk_gm_008",
-    "chunk_gm_009"
+    "chunk_gm_001a",
+    "chunk_gm_001b",
+    "chunk_gm_001c",
+    "chunk_gm_002a",
+    "chunk_gm_002b",
+    "chunk_gm_003a",
+    "chunk_gm_003b",
+    "chunk_gm_004a",
+    "chunk_gm_004b",
+    "chunk_gm_005a",
+    "chunk_gm_005b",
+    "chunk_gm_006a",
+    "chunk_gm_006b",
+    "chunk_gm_007a",
+    "chunk_gm_007b",
+    "chunk_gm_008a",
+    "chunk_gm_008b",
+    "chunk_gm_009a",
+    "chunk_gm_009b"
   ],
   "visibilityConditionId": "cond_always",
   "accessConditionId": "cond_always",
