+++ content/gameplay/transmissions/tx_where_are_you.json
@@ -13,16 +13,28 @@
     "evacuation"
   ],
   "transcriptChunkIds": [
-    "chunk_way_001",
-    "chunk_way_002",
-    "chunk_way_003",
-    "chunk_way_004",
-    "chunk_way_005",
-    "chunk_way_006",
-    "chunk_way_007",
-    "chunk_way_008",
-    "chunk_way_009",
-    "chunk_way_010"
+    "chunk_way_001a",
+    "chunk_way_001b",
+    "chunk_way_001c",
+    "chunk_way_002a",
+    "chunk_way_002b",
+    "chunk_way_002c",
+    "chunk_way_003a",
+    "chunk_way_003b",
+    "chunk_way_004a",
+    "chunk_way_004b",
+    "chunk_way_005a",
+    "chunk_way_005b",
+    "chunk_way_006a",
+    "chunk_way_006b",
+    "chunk_way_007a",
+    "chunk_way_007b",
+    "chunk_way_008a",
+    "chunk_way_008b",
+    "chunk_way_009a",
+    "chunk_way_009b",
+    "chunk_way_010a",
+    "chunk_way_010b"
   ],
   "visibilityConditionId": "cond_mission_good_morning_cleared",
   "accessConditionId": "cond_mission_good_morning_cleared",
