+++ docs/02_開発体制とディレクトリ構成.md
@@ -1,6 +1,6 @@
 # 02_開発体制とディレクトリ構成
 
-更新日: 2026-04-17
+更新日: 2026-04-26
 
 ## この文書の目的
 
@@ -133,13 +133,20 @@ IndexedDB
 
 主なファイル:
 - `packages/game-session/src/game-session.ts`
+- `packages/game-session/src/battle-effects.ts`
+- `packages/game-session/src/battle-state.ts`
+- `packages/game-session/src/battle-world.ts`
+- `packages/game-session/src/explore-world.ts`
 - `packages/game-session/src/equipment-runtime.ts`
+- `packages/game-session/src/profile-factory.ts`
 - `packages/game-session/src/progression.ts`
+- `packages/game-session/src/session-snapshots.ts`
 - `packages/game-session/src/selectors.ts`
 - `packages/game-session/src/conditions.ts`
 - `packages/game-session/src/hazards.ts`
 - `packages/game-session/src/presentation.ts`
 - `packages/game-session/src/runtime-types.ts`
+- `packages/game-session/src/record-utils.ts`
 
 責務の分け方:
 - `game-session.ts`
@@ -147,13 +154,27 @@ IndexedDB
   - stepExplore / stepBattle
   - save 更新
   - mission 開始と終了
+- `battle-effects.ts`
+  - `RuntimeEffectRequest` の戦闘状態への適用
+  - projectile 生成、弾の移動、爆発処理
+- `battle-state.ts`
+  - session 内部で使う戦闘状態の型
+- `battle-world.ts`
+  - 戦闘フィールド寸法、spawn、当たり判定、弾幕方向などの純粋計算
+- `explore-world.ts`
+  - 探索範囲、マップ境界、reveal bitmap、探索移動、近傍 node 判定
 - `equipment-runtime.ts`
   - 装備の解決
   - 強化値の適用
   - main / sub / subsystem hook の handler 解決
+- `profile-factory.ts`
+  - 新規 profile / save slot / 初期探索履歴の生成
 - `progression.ts`
   - 解析率、復元率、エリア解放率、通信状態の共通計算
   - UI と session で同じ式を使うための pure function
+- `session-snapshots.ts`
+  - profile / battle state から snapshot を組み立てる pure function
+  - active profile がない時は空 snapshot を返し、仮 profile は生成しません
 - `selectors.ts`
   - 未開放情報を UI に渡さないための絞り込み
   - feature access
@@ -203,8 +224,10 @@ IndexedDB
 責務の分け方:
 - `use-magnolia-app.ts`
   - session の初期化
-  - input を command / frame input に変換
   - session から snapshot / renderState を取り出して React state へ反映
+- `use-magnolia-input.ts`
+  - keyboard / mouse state の保持
+  - input を command / frame input に変換するための読み取り
 - `App.tsx`
   - screen ごとの分岐だけ
 - `display-helpers.ts`
@@ -220,8 +243,12 @@ IndexedDB
   - タイトルとセーブスロット選択を同一画面で扱います
 - `screens/*`
   - 画面単位の UI 合成
+  - `screens/menu/MenuScreen.tsx` は tab 合成だけを持ち、装備・アーカイブ・設定は同階層の panel に分けます
 - `components/*`
   - Canvas 描画や再利用 UI
+  - `components/BattleCanvas.tsx` は canvas の接続だけを担当し、戦闘描画本体は `components/battle-renderer.ts` に置きます
+- `styles/global.css`
+  - CSS の読み込み入口です。実体は `styles/base.css`、画面単位 CSS、`styles/responsive.css` へ分けます
 
 ## フロントエンド改修の入口
