+++ docs/05_データ構造.md
@@ -24,8 +24,12 @@
   - 通信に接続した後の戦闘単位です。時間、敵配置、字幕進行、解析対象量を持ちます。
 - `heardRanges`
   - プレイヤーが欠損させずに確保できた時間範囲です。本文開放はこの配列の累積で決めます。
+  - ミッション中に過去の復元済み区間を seed として含む場合があります。
 - `damageRanges`
   - 被弾やノイズで欠損した時間範囲です。ラン単位の記録に残します。
+- `seededHeardRanges`
+  - 再挑戦前にすでに復元済みだった時間範囲です。
+  - 今回の被弾で過去ログを失わないため、今回分の復元計算から分けて扱います。
 - スコア
   - 1 ラン内のパフォーマンス指標です。解析率・復元率・行動から算出します。消費しません。
 - 自己修復ポイント
@@ -58,6 +62,8 @@
 
 運用ルール:
 - 仮本文は短い placeholder でもよいですが、時間範囲だけは実際の構成に近い粒度で切ります。
+- アーカイブ本文は完全に復元できた chunk だけを表示するため、本文チャンクは 1〜4 秒程度を目安に、文や意味の切れ目で分けます。
+- 被弾した時間範囲は今回の復元対象から除外します。過去に確保済みの `seededHeardRanges` は失いません。
 - 仮説明文は後で差し替えてよいですが、ID と参照関係は変更しない前提で進めます。
 - 値が未定でも `undefined` で逃がさず、仮値を入れて validation を通します。
