+++ docs/08_表示と機能の境界.md
@@ -1,6 +1,6 @@
 # 08_表示と機能の境界
 
-更新日: 2026-04-17
+更新日: 2026-04-26
 
 ## この文書の目的
 
@@ -50,6 +50,7 @@
 2. 表示に必要な値が足りない場合
    - `runtime-types.ts` に renderState を追加します。
    - `game-session.ts` で renderState を組み立てます。
+   - snapshot の組み立てだけで済む場合は `session-snapshots.ts` を更新します。
    - UI はその追加フィールドだけを読みます。
 3. gameplay 調整の場合
    - まず `content/gameplay/*.json` で表現できるか確認します。
@@ -73,3 +74,12 @@
 - settings 変更は `changeSetting` command 経由でのみ行います。
 - 装備購入、強化、装着、アーカイブ選択、mission 開始は session 側で検証します。
 - `saveToSlot` は「別スロットへ複製して保存」です。現在操作中のスロットは切り替えません。
+
+## 現行の画面・描画境界
+
+- `MenuScreen.tsx` は tab と画面合成だけを持ちます。
+- 装備、アーカイブ、設定の各表示は `apps/web/src/screens/menu/*Panel.tsx` に置きます。
+- `BattleCanvas.tsx` は canvas 要素の初期化だけを持ちます。
+- 戦闘の描画順、描画 registry、個別の描画関数は `apps/web/src/components/battle-renderer.ts` に置きます。
+- 戦闘の effect 適用、弾生成、弾移動、爆発処理は `packages/game-session/src/battle-effects.ts` に置きます。
+- CSS の入口は `apps/web/src/styles/global.css` です。実装は画面単位の CSS ファイルへ分け、変更対象の画面に対応するファイルを編集します。
